fix: Refactor Login and Signup components
