+++ src/components/SignInUp/CommonContent/index.ts
@@ -1,0 +1,1 @@
+export * from './CommonContent.styled';
