+++ src/components/SignInUp/SignNavigation/SignNavigation.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
-import { BackButton } from '../BackButton';
-import { Logo } from '../Logo';
+import { BackButton } from '../../BackButton';
+import { Logo } from '../../Logo';
 import { SignNavigationStyled } from './SignNavigation.styled';
 
 export const SignNavigation: FC = () => {
