+++ src/pages/Login/Login.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
 import { FC } from 'react';
 import { useMediaQueries } from '../../hooks/useMediaQueries';
-import { LoginPhones } from '../../components/LoginPhones';
-import { LoginDesktops } from '../../components/LoginDesktops';
+import { LoginPhones } from '../../components/SignInUp/LoginPhones';
+import { LoginDesktops } from '../../components/SignInUp/LoginDesktops';
 
 export const Login: FC = () => {
   const { md } = useMediaQueries();
