+++ src/pages/Signup/Signup.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import { useMediaQueries } from '../../hooks/useMediaQueries';
-import { SignupPhones } from '../../components/SignupPhones';
-import { SignupDesktops } from '../../components/SignupDesktops';
+import { SignupPhones } from '../../components/SignInUp/SignupPhones';
+import { SignupDesktops } from '../../components/SignInUp/SignupDesktops';
 
 export const Signup: FC = () => {
   const { md } = useMediaQueries();
